fixed readme and code
