+++ README.md
@@ -1,18 +1,2 @@
-# Text Encrypter
-## What is this?
-Text Encrypter CLUI made with C++        
-Uses [Symmetric](https://en.wikipedia.org/wiki/Symmetric-key_algorithm) 
-[ASCII Encryption](https://www.google.com/search?q=ASCII+Encryption&rlz=1C1JJTC_enIN929IN929&sourceid=chrome&ie=UTF-8)
-technique and [File Format Encryption](https://github.com/thenithinbalaji/Text-Encrypter/blob/main/README.md#file-formats) Technique to encrypt the user data.
-
-## Usage
-+ Download and Run [text-encrypter.cpp](text-encrypter.cpp)
-+ When the Encryption option is chosen, the user is asked to enter the filename. 
-   + After the file is chosen, the app uses its in-built algorithm to encrypt the file and stores it in filename.en format. 
-   + The filename.en is completely unreadable by any software except this file encryptor program. 
-+ Similarly, while decrypting, the filename.en file is chosen and the data inside it is decrypted back to text format. 
- 
-## File Formats
-`filename.txt`    
-`filename.en`
-
+Text Encrypter CLUI made with C++.   
+Uses [Symmetric](https://en.wikipedia.org/wiki/Symmetric-key_algorithm) [ASCII Encryption](https://www.google.com/search?q=ASCII+Encryption&rlz=1C1JJTC_enIN929IN929&sourceid=chrome&ie=UTF-8) and [File Format Encryption](https://github.com/thenithinbalaji/Text-Encrypter/blob/main/README.md#file-formats) technique to encrypt the user data.
